+++ apps/alchemind/CHANGELOG.md
@@ -11,5 +11,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial implementation of core behaviors
 - Application structure
 - Basic integration tests
+- Added speech callback and tts helper function
 
 [Unreleased]: https://github.com/bradleygolden/alchemind/compare/HEAD...HEAD